Add placeholder Foe P1 npc class and level actors spawning
The positions and entity (actor) types are being read from the game
level file and created on the map.

Atm a default NPC FoeP1 entity class exists. This should be further
developed to support AI, movement and fight routines. We would need some
sort of a base class to impl. common routines for all NPCs.